chore: loader styles in module, move some files
