+++ index.html
@@ -7,7 +7,6 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>6 cities</title>
   <link rel="stylesheet" href="css/main.css">
-  <link rel="stylesheet" href="css/loader.css">
 </head>
 
 <body>
