+++ src/components/loader/loader.module.css
@@ -1,4 +1,4 @@
-.loader-box {
+.loader_box {
   width: 100%;
   height: 100%;
   position: relative;
