+++ src/components/loader/loader.tsx
@@ -1,11 +1,13 @@
+import styles from './loader.module.css';
+
 type Props = {
   size?: number;
 }
 
 export default function Loader({ size }: Props): JSX.Element {
   return (
-    <div className="loader-box">
-      <span className="loader" style={size ? { width: `${size}px`, height: `${size}px` } : undefined} ></span>
+    <div className={styles.loader_box}>
+      <span className={styles.loader} style={size ? { width: `${size}px`, height: `${size}px` } : undefined} ></span>
     </div>
   );
 }
