+++ src/store/auth.ts
@@ -1,13 +1,13 @@
 import { createAction, createAsyncThunk, createSlice } from '@reduxjs/toolkit';
-import { AuthorizationStatus } from '../../models/authorization-status';
-import { UserModel } from '../../models/user-model';
-import { Namespace } from '../namespace';
-import handleRequest from '../../tools/handle-request';
-import { dropToken, saveToken } from '../../services/token';
-import { RegistrationRequest } from '../../models/registration-request';
+import { AuthorizationStatus } from '../models/authorization-status';
+import { UserModel } from '../models/user-model';
+import { Namespace } from './namespace';
+import handleRequest from '../tools/handle-request';
+import { dropToken, saveToken } from '../services/token';
+import { RegistrationRequest } from '../models/registration-request';
 import { AxiosInstance } from 'axios';
 import { fetchFavoriteOffers } from './offers';
-import { ApiRoute } from '../../configuration/api-route';
+import { ApiRoute } from '../configuration/api-route';
 
 export type AuthState = {
   authStatus: AuthorizationStatus;
