+++ src/store/reducer.ts
@@ -1,7 +1,7 @@
 import { combineReducers } from '@reduxjs/toolkit';
 import { Namespace } from './namespace';
-import { authSlice } from './namespaces/auth';
-import { offersSlice } from './namespaces/offers';
+import { authSlice } from './auth';
+import { offersSlice } from './offers';
 
 export const rootReducer = combineReducers({
   [Namespace.Auth]: authSlice.reducer,
